+++ CHANGELOG.md
@@ -9,10 +9,43 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- **GM-781** - [HOME] Contentful images optimization
+
+### Fixed
+
+- **GM-845** - [LOGIN] Password creation regex and verification
+- **GM-827** - fix orders to use the new API data
+- **GM-865** - Adjust products endpoint
+- **GM-675** - Paginator go to last page
+- **GM-784** - [MENU] Adjust in banner menu filters
+- **GM-611** - Change product description component 
+
+### Removed
+
+## [0.20.0] - 2022-09-29
+
+### Added
+
+- **GM-796** - Integration with GoogleAnalytics 4 and Partytown
+
+### Fixed
+
+- **GM-672** - Adjust arrows Carousel in main banner
+- **GM-820** - [MENU] Adjust in margin of popouver
+- **GM-674** - [HOME] Adjust in video section for cursor pointer
+- **GM-274** - Resolved bug Common Questions
+
+
+### Removed
+
+## [0.19.0] - 2022-09-22
+
+### Added
+
 - **GM-330** - Login integration
 - **GM-742** - Page My Courses and integrations with truechange
 
-### Fixed 
+### Fixed
 
 - **GM-783** - [Planos] Adjust select and options
 - **GM-495** - Adjust banner in infoprodutor page
@@ -23,7 +56,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **GM-502** - [PDP] Adjust of redirection to plans page
 
 ### Removed
-
 
 ## [0.18.0] - 2022-09-20
 
@@ -42,7 +74,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **GM-349** - [HOME] Adjusted Newletter layout on mobile
 - **GM-657** - [Search] Adjusted Search input layout and behaviour
 - **GM-794** - [PDP] Remove lowercase of linkId in shelfs and search results
-
 
 ## [0.17.1] - 2022-09-15